--- a/server/grammar/comments.bnf
+++ b/server/grammar/comments.bnf
@@ -1,4 +1,4 @@
 score         = (non-comment | <comment> | clj-expr)*
 comment       = <"#"> #".*" <#"(\n|\r|$)+">
-<non-comment> = !("#" | "(" | ")") #".|\n|\r"
+<non-comment> = #"[^\#\(\)]+"