--- a/server/grammar/duration.bnf
+++ b/server/grammar/duration.bnf
@@ -1,8 +1,8 @@
 duration                = (note-length | seconds | milliseconds)
-                          <ows> barline? <ows> subduration* slur?
+                          (<ows> barline? <ows> subduration)*
+                          (barline? <ows> slur)?
 <subduration>           = tie <ows> barline? <ows>
                           (note-length | seconds | milliseconds) <ows>
-                          barline? <ows>
 
 seconds                 = positive-number <"s">
 milliseconds            = positive-number <"ms">