I made a pass through each of Alda's parsers with `instaparse.core/parses` to determine whether there was any ambiguity contributing to slow parse times noted in #208. I did find a few places where there was ambiguity in the grammar -- see the commit messages for more details.

Unfortunately, this is only a minor improvement over what we have now. The major example of a slow-parsing score we have now is the Bach cello suite example. Currently on my computer, that score takes about 800-900 ms to parse. With the improvements on this branch, it takes about 700-800 ms which is barely a difference.

I think the next step is going to be considering some of the other suggestions in the [instaparse wiki's performance tips](https://github.com/Engelberg/instaparse/blob/master/docs/Performance.md#performance-tips). This PR should make performance tip 5 no longer an issue, but maybe performance tip 6 is applicable.

One thing I started playing around with was commenting out certain parts of the parsing pipeline and using `clojure.core/time` to profile the different stages. On my computer, it took about 300 ms just to remove the comments and parse Clojure expressions, which seems fishy.

(WIP PR -- will leave open, at least for now. I'd love some feedback, if anyone reading this happens to be knowledgeable about parsing and/or can help determine the causes of Alda's slowness in parsing larger scores.)