@daveyarwood 2016-06-04
@aengelberg Do you have any ideas about what might be causing the comment parser to take 300 ms on the Bach cello suite example?

The comment parser is a concatenation of the [comments.bnf](https://github.com/alda-lang/alda/blob/master/server/grammar/comments.bnf) and [clojure.bnf](https://github.com/alda-lang/alda/blob/master/server/grammar/clojure.bnf) grammars.

I ran the score through `(insta/parses comment-parser :partial true)` and there are a ton of steps where it adds `music-data` one character at a time. Could that be part of the issue?