@aengelberg 2016-06-04
Out of the 555ms (on my machine) it took to parse the Bach cello suite, about 125ms of that was spent removing comments. I was able to at least shave that off by optimizing the comment parser like so:
```
<non-comment> = #"[^\#\(\)]+"
```
```clojure
;; before change
alda.parser> (time (def parsed (comment-parser (slurp "examples/bach_cello_suite_no_1.alda"))))
"Elapsed time: 126.040464 msecs"
;; after change
alda.parser> (time (def parsed (comment-parser (slurp "examples/bach_cello_suite_no_1.alda"))))
"Elapsed time: 7.186078 msecs"
```