@aengelberg 2016-06-04
I was able to take the `score-parser` from 120ms down to 3ms with the following optimization to `score.bnf`:
```
<music-data>      = !calls #"[a-zA-Z]*[^a-zA-Z]*([a-zA-Z][^a-zA-Z]+)*"
```
This takes some thought to fully verify it's doing the right thing. My reasoning is:
- With `!calls`, we guarantee a new `part` is not starting at the current character.
- A `part` always starts with a `name` which would begin with at least two letters.
- Thus, we can safely eat all of the next letters in a row, because if a `part` started after some but not all of the next letters, those letters would have to be `part` of the part name. Hence `[a-zA-Z]*`.
- Now, we can safely eat all of the next non-letters in a row (i.e. numbers, special characters, and whitespace), because a `part` may not start at the current point or in between any of those non-letters. Hence `[^a-zA-Z]*`.
- At this point, we can now eat any letter not followed by another letter, because an `acceptable-name` starts with at least two letters in a row. So we finish the regex off with `([a-zA-Z][^a-zA-Z]+)*`.

Please feel free to correct me if that proof isn't airtight.

A well-known issue with Java regexes would impact this approach: the regex engine will stack-overflow when trying to match thousands of instances of a grouped (`(...)`) expression. Changing to `(?:...)` does not help. Not sure if Alda scores in practice would be affected by this, but that's a definite issue.

The `part-parser` appears to be the primary bottleneck (288ms), so I'll report back if I find any optimizations to that.