@daveyarwood 2016-06-05
This helps a lot! These 2 changes alone already makes a huge difference.

It took me a while to wrap my head around your `music-data` regex, but I think I'm getting it now. If I'm understanding correctly, the idea is: instead of reading the part's music-data one character at a time, we read it in chunks, alternating between letters, non-letters, and "one-letter-followed-by-non-letters," and at the end of each pass through these 3 things, we check that the next thing read isn't a `calls`, then repeat. Do I have that right? At any rate, it definitely seems to do the trick! 👍 

It's good to know about the Java regex stack overflow issue... it will be interesting to see if this becomes an issue. I tried to trigger it by making a really long instrument part -- copy-pasting the cello part in the Bach example onto the end of itself -- but ran into "Method code too large!" (see issue #235 ) first. I'll do some more testing once I'm able to resolve that issue, but I think as a general rule, it will be fairly rare for a single instrument part in an Alda score to get big enough to trigger "Method code too large!" on its own, let alone to trigger the regex group stack overflow. I think in the near future, the recommended score-writing method is going to be to define a bunch of sequences of notes (outside of parts) and then compose the instrument parts out of the defined sequences by referencing variables ( #131 ), so the use case for having a super long instrument part will become less and less.

I really appreciate all the help! I'm excited to see if you find anything else we can optimize in `part-parser`.